+++ src/lang/program.cpp
@@ -14,7 +14,7 @@ const std::array<Operation::Type, 31> Operation::Types = {
     Operation::Type::LEQ, Operation::Type::GEQ, Operation::Type::MIN,
     Operation::Type::MAX, Operation::Type::BAN, Operation::Type::BOR,
     Operation::Type::BXO, Operation::Type::LPB, Operation::Type::LPE,
-    Operation::Type::CLR, Operation::Type::SOR, Operation::Type::SEQ,
+    Operation::Type::CLR, Operation::Type::SRT, Operation::Type::SEQ,
     Operation::Type::DBG,
 };
 
@@ -75,8 +75,8 @@ const Operation::Metadata& Operation::Metadata::get(Type t) {
       Operation::Type::LPE, "lpe", 0, true, false, false};
   static Operation::Metadata clr{
       Operation::Type::CLR, "clr", 2, true, false, true};
-  static Operation::Metadata sor{
-      Operation::Type::SOR, "sor", 2, true, true, true};
+  static Operation::Metadata srt{
+      Operation::Type::SRT, "srt", 2, true, true, true};
   static Operation::Metadata seq{
       Operation::Type::SEQ, "seq", 2, true, true, true};
   static Operation::Metadata dbg{
@@ -138,8 +138,8 @@ const Operation::Metadata& Operation::Metadata::get(Type t) {
       return lpe;
     case Operation::Type::CLR:
       return clr;
-    case Operation::Type::SOR:
-      return sor;
+    case Operation::Type::SRT:
+      return srt;
     case Operation::Type::SEQ:
       return seq;
     case Operation::Type::DBG: